Add Fenix feature annotations, based on issue metadata (#12)
In many cases, is possible to infer the feature a fenix metric corresponds
to from the issue metadata corresponding to the data review. Scrape this
data and add it to a prototype "labels" field.

This PR also adds a schema for creating labels for metrics. Currently
this is Fenix-only and is subject to change without warning.

Co-authored-by: Jeff Klukas <jklukas@mozilla.com>