Allow "content" for an annotation to be blank and remove stub commentary
It's better just to leave this blank, we can get the Glean Dictionary
to add some kind of placeholder when we have (for example) tags but no
annotations, which is increasingly common for Fenix.

This also changes the API slightly to change the "content" property
for an annotation to "commentary" (makes consuming this API inside
the Glean Dictionary slightly easier).